Added the 'conn' and 'Dashboard' class.
The conn class is created but doesn't work yet due to some error in jdbc
connection. This needs to be fixed.

The Dashboard class is also created and works individually but not in sync
with the Login class. This will be fixed when conn is available.

Afew more files including images are also added in the assets.